# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Documentation for Product API",
    "title": "of Product API",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
    "/products": {
      "get": {
        "description": "Returns a list of products",
        "tags": [
          "products"
        ],
        "operationId": "listProducts",
        "responses": {
          "200": {
            "$ref": "#/responses/productsResponse"
          }
        }
      }
    },
    "/products/{id}": {
      "put": {
        "description": "Returns a list of products",
        "tags": [
          "products"
        ],
        "operationId": "updateProducts",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The id of the product to update from the database",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/noContent"
          }
        }
      }
    }
  },
  "definitions": {
    "Product": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "id": {
          "description": "The id of this user",
          "type": "integer",
          "format": "int64",
          "minimum": 1,
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "price": {
          "type": "number",
          "format": "float",
          "x-go-name": "Price"
        },
        "sku": {
          "type": "string",
          "x-go-name": "SKU"
        }
      },
      "x-go-package": "example.com/data"
    }
  },
  "responses": {
    "noContent": {
      "description": ""
    },
    "productsResponse": {
      "description": "A list of products returns in the response",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Product"
        }
      }
    }
  }
}